Revert "feat: Tai: thiết kế giao diện đăng nhập, điểm số,, học sinh (#4)"
This reverts commit 88204c1f4b6d1c66b008abd1f107de16815e709a.
