+++ Student-Management-System_CSharp_SGU2025/GUI/FrmDangNhap.cs
@@ -1,0 +1,45 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.Data;
+using System.Drawing;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows.Forms;
+
+namespace Student_Management_System_CSharp_SGU2025.GUI
+{
+    public partial class FrmDangNhap : Form
+    {
+        public FrmDangNhap()
+        {
+            InitializeComponent();
+        }
+
+        private void label1_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void label2_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void label3_Click(object sender, EventArgs e)
+        {
+
+        }
+
+        private void FrmDangNhap_Load(object sender, EventArgs e)
+        {
+
+        }
+
+        private void lbTenDangNhap_Click(object sender, EventArgs e)
+        {
+
+        }
+    }
+}
