--- Student-Management-System_CSharp_SGU2025/GUI/TaiSuDung/HeaderQuanLiHocSinh.cs
@@ -1,26 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace Student_Management_System_CSharp_SGU2025.GUI
-{
-    public partial class HeaderQuanLiHocSinh : UserControl
-    {
-        
-        public HeaderQuanLiHocSinh()
-        {
-            InitializeComponent();
-        }
-
-        private void HeaderQuanLiHocSinh_Load(object sender, EventArgs e)
-        {
-
-        }
-    }
-}
