--- Student-Management-System_CSharp_SGU2025/GUI/TaiSuDung/StatCardQuanLiHocSinh.cs
@@ -1,31 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Data;
-using System.Drawing;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Forms;
-
-namespace Student_Management_System_CSharp_SGU2025.GUI
-{
-    public partial class StatCardQuanLiHocSinh : UserControl
-    {
-        public StatCardQuanLiHocSinh()
-        {
-            InitializeComponent();
-
-        }
-
-        private void StatCardQuanLiHocSinh_Load(object sender, EventArgs e)
-        {
-
-        }
-
-        private void lbCardValue_Click(object sender, EventArgs e)
-        {
-
-        }
-    }
-}
