+++ Student-Management-System_CSharp_SGU2025/Program.cs
@@ -13,13 +13,7 @@ namespace Student_Management_System_CSharp_SGU2025
         {
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
-            // Fix: Application.Run requires a Form, not a UserControl.
-            // Wrap HocSinh (UserControl) in a Form before running.
-            Form mainForm = new Form();
-            GUI.HanhKiem hanhKiem = new GUI.HanhKiem();
-            hanhKiem.Dock = DockStyle.Fill;
-            mainForm.Controls.Add(hanhKiem);
-            Application.Run(mainForm);
+            Application.Run(new GUI.FrmDangNhap());
         }
     }
 }
