+++ Student-Management-System_CSharp_SGU2025/Student-Management-System_CSharp_SGU2025.csproj
@@ -71,35 +71,11 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="DTO\MonHoc.cs" />
-    <Compile Include="GUI\DiemSo\DiemSo_NhapDiem.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="GUI\DiemSo\DiemSo_NhapDiem.Designer.cs">
-      <DependentUpon>DiemSo_NhapDiem.cs</DependentUpon>
-    </Compile>
-    <Compile Include="GUI\DangNhap\FrmDangNhap.cs">
+    <Compile Include="GUI\FrmDangNhap.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="GUI\DangNhap\FrmDangNhap.Designer.cs">
+    <Compile Include="GUI\FrmDangNhap.Designer.cs">
       <DependentUpon>FrmDangNhap.cs</DependentUpon>
-    </Compile>
-    <Compile Include="GUI\LamThem\HanhKiem.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="GUI\LamThem\HanhKiem.Designer.cs">
-      <DependentUpon>HanhKiem.cs</DependentUpon>
-    </Compile>
-    <Compile Include="GUI\TaiSuDung\HeaderQuanLiHocSinh.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="GUI\TaiSuDung\HeaderQuanLiHocSinh.Designer.cs">
-      <DependentUpon>HeaderQuanLiHocSinh.cs</DependentUpon>
-    </Compile>
-    <Compile Include="GUI\HocSinh\HocSinh.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="GUI\HocSinh\HocSinh.Designer.cs">
-      <DependentUpon>HocSinh.cs</DependentUpon>
     </Compile>
     <Compile Include="GUI\RecentActivityItem.cs">
       <SubType>UserControl</SubType>
@@ -125,12 +101,6 @@
     <Compile Include="GUI\FrmMonHoc.Designer.cs">
       <DependentUpon>FrmMonHoc.cs</DependentUpon>
     </Compile>
-    <Compile Include="GUI\TaiSuDung\StatCardQuanLiHocSinh.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Include="GUI\TaiSuDung\StatCardQuanLiHocSinh.Designer.cs">
-      <DependentUpon>StatCardQuanLiHocSinh.cs</DependentUpon>
-    </Compile>
     <Compile Include="GUI\TrangChu.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -139,23 +109,8 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <EmbeddedResource Include="GUI\DiemSo\DiemSo_NhapDiem.resx">
-      <DependentUpon>DiemSo_NhapDiem.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GUI\DangNhap\FrmDangNhap.resx">
+    <EmbeddedResource Include="GUI\FrmDangNhap.resx">
       <DependentUpon>FrmDangNhap.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GUI\LamThem\HanhKiem.resx">
-      <DependentUpon>HanhKiem.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GUI\TaiSuDung\HeaderQuanLiHocSinh.resx">
-      <DependentUpon>HeaderQuanLiHocSinh.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GUI\HocSinh\HocSinh.resx">
-      <DependentUpon>HocSinh.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Include="GUI\TaiSuDung\StatCardQuanLiHocSinh.resx">
-      <DependentUpon>StatCardQuanLiHocSinh.cs</DependentUpon>
     </EmbeddedResource>
     <EmbeddedResource Include="GUI\TrangChu.resx">
       <DependentUpon>TrangChu.cs</DependentUpon>
@@ -190,14 +145,11 @@
   </ItemGroup>
   <ItemGroup>
     <Content Include="Images\3que.png" />
-    <None Include="Images\download.png" />
-    <None Include="Images\diskette.png" />
     <Content Include="Images\education.png">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
     <None Include="Images\people.png" />
     <None Include="Images\padlock.png" />
-    <None Include="Images\plus.png" />
     <Content Include="Images\repair.png" />
     <None Include="Images\school.png" />
   </ItemGroup>
